UNIONVMS-747 refacoring
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Reporting/APP@5793 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/reporting/service/dto/SegmentDTO.java
@@ -72,10 +72,10 @@ public class SegmentDTO {
 
     private interface Include {
         String getId();
-        BigDecimal getCourseOverGround();
-        BigDecimal getSpeedOverGround();
-        BigDecimal getDuration();
-        BigDecimal getDistance();
+        Double getCourseOverGround();
+        Double getSpeedOverGround();
+        Double getDuration();
+        Double getDistance();
         String getTrackId();
     }
